Extracting logstash metricset into own file
