+++ metricbeat/module/logstash/metricset.go
@@ -1,0 +1,51 @@
+// Licensed to Elasticsearch B.V. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Elasticsearch B.V. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package logstash
+
+import (
+	"github.com/elastic/beats/v7/metricbeat/helper"
+	"github.com/elastic/beats/v7/metricbeat/helper/elastic"
+	"github.com/elastic/beats/v7/metricbeat/mb"
+)
+
+// MetricSet can be used to build other metricsets within the Logstash module.
+type MetricSet struct {
+	mb.BaseMetricSet
+	*helper.HTTP
+	XPack bool
+}
+
+// NewMetricSet creates a metricset that can be used to build other metricsets
+// within the Logstash module.
+func NewMetricSet(base mb.BaseMetricSet) (*MetricSet, error) {
+	config := elastic.DefaultModeConfig()
+	if err := base.Module().UnpackConfig(&config); err != nil {
+		return nil, err
+	}
+
+	http, err := helper.NewHTTP(base)
+	if err != nil {
+		return nil, err
+	}
+
+	return &MetricSet{
+		base,
+		http,
+		config.GetMode() == elastic.ModeStackMonitoring,
+	}, nil
+}
